Add: post Article success
