+++ components/boards/BestContent.tsx
@@ -1,4 +1,4 @@
-import { Article } from '@/lib/axios';
+import { Article } from '@/lib/api';
 import Image from 'next/image';
 
 type Props = {
